Continued work on boss and boss fog
Improved how boss fog works, so it should mostly look good now and fade out to the natural fog conditions no matter when or where it is, except when a boss dies. If that happens it will snap to the current fog conditions (wip).
Also added a new bat swarm attack on boss zombies.